requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pdfminer.six==20231228
streamlit==1.38.0
google-generativeai>=0.7.0
